--- packages/trustix/client/unixtransport/go.mod
@@ -1,3 +1,0 @@
-module github.com/nix-community/trustix/packages/trustix/client/unixtransport
-
-go 1.16
