+++ packages/unixtransport/go.mod
@@ -1,0 +1,3 @@
+module github.com/nix-community/trustix/packages/unixtransport
+
+go 1.16
